+++ src/Telegram.Bot/ITelegramBotClient.cs
@@ -381,7 +381,7 @@ namespace Telegram.Bot
         /// <see href="https://core.telegram.org/bots/api#sendvideo"/>
         Task<Message> SendVideoAsync(
             ChatId chatId,
-            FileToSend video,
+            InputOnlineFile video,
             int duration = default,
             int width = default,
             int height = default,
@@ -429,7 +429,7 @@ namespace Telegram.Bot
         /// <see href="https://core.telegram.org/bots/api#sendvideonote"/>
         Task<Message> SendVideoNoteAsync(
             ChatId chatId,
-            FileToSend videoNote,
+            InputTelegramFile videoNote,
             int duration = default,
             int length = default,
             bool disableNotification = default,