+++ src/Telegram.Bot/Requests/SendVideoNoteRequest.cs
@@ -1,10 +1,10 @@
-﻿using System.Collections.Generic;
-using System.Net.Http;
+﻿using System.Net.Http;
 using Newtonsoft.Json;
-using Telegram.Bot.Helpers;
+using Newtonsoft.Json.Serialization;
 using Telegram.Bot.Requests.Abstractions;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
+using Telegram.Bot.Types.InputFiles;
 using Telegram.Bot.Types.ReplyMarkups;
 
 namespace Telegram.Bot.Requests
@@ -12,6 +12,7 @@ namespace Telegram.Bot.Requests
     /// <summary>
     /// Send rounded video messages
     /// </summary>
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
     public class SendVideoNoteRequest : FileRequestBase<Message>,
                                         INotifiableMessage,
                                         IReplyMessage,
@@ -20,12 +21,14 @@ namespace Telegram.Bot.Requests
         /// <summary>
         /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
         /// </summary>
+        [JsonProperty(Required = Required.Always)]
         public ChatId ChatId { get; set; }
 
         /// <summary>
         /// Video note to send
         /// </summary>
-        public FileToSend VideoNote { get; set; }
+        [JsonProperty(Required = Required.Always)]
+        public InputTelegramFile VideoNote { get; set; }
 
         /// <summary>
         /// Duration of sent video in seconds
@@ -63,58 +66,17 @@ namespace Telegram.Bot.Requests
         /// </summary>
         /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
         /// <param name="videoNote">Video note to send</param>
-        public SendVideoNoteRequest(ChatId chatId, FileToSend videoNote)
+        public SendVideoNoteRequest(ChatId chatId, InputTelegramFile videoNote)
             : this()
         {
             ChatId = chatId;
             VideoNote = videoNote;
         }
 
-        /// <summary>
-        /// Generate content of HTTP message
-        /// </summary>
-        /// <returns>Content of HTTP request</returns>
-        public override HttpContent ToHttpContent()
-        {
-            HttpContent content;
-
-            if (VideoNote.Type == FileType.Stream)
-            {
-                var parameters = new Dictionary<string, object>
-                {
-                    { nameof(ChatId).ToSnakeCased(), ChatId},
-                    { nameof(VideoNote).ToSnakeCased(), VideoNote},
-                    { nameof(ReplyMarkup).ToSnakeCased(), ReplyMarkup }
-                };
-
-                if (Duration != default)
-                {
-                    parameters.Add(nameof(Duration).ToSnakeCased(), Duration);
-                }
-
-                if (Length != default)
-                {
-                    parameters.Add(nameof(Length).ToSnakeCased(), Length);
-                }
-
-                if (ReplyToMessageId != default)
-                {
-                    parameters.Add(nameof(ReplyToMessageId).ToSnakeCased(), ReplyToMessageId);
-                }
-
-                if (DisableNotification != default)
-                {
-                    parameters.Add(nameof(DisableNotification).ToSnakeCased(), DisableNotification);
-                }
-
-                content = GetMultipartContent(parameters);
-            }
-            else
-            {
-                content = base.ToHttpContent();
-            }
-
-            return content;
-        }
+        /// <inheritdoc />
+        public override HttpContent ToHttpContent() =>
+            VideoNote.FileType == FileType.Stream
+                ? ToMultipartFormDataContent("video_note", VideoNote)
+                : base.ToHttpContent();
     }
 }