+++ src/Telegram.Bot/Types/File.cs
@@ -7,8 +7,7 @@ namespace Telegram.Bot.Types
     /// <summary>
     /// This object represents a file ready to be downloaded. The file can be downloaded via <see cref="TelegramBotClient.GetFileAsync"/>. It is guaranteed that the link will be valid for at least 1 hour. When the link expires, a new one can be requested by calling <see cref="TelegramBotClient.GetFileAsync"/>.
     /// </summary>
-    [JsonObject(MemberSerialization = MemberSerialization.OptIn,
-                NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
     public class File
     {
         /// <summary>
@@ -20,13 +19,13 @@ namespace Telegram.Bot.Types
         /// <summary>
         /// Optional. File size, if known
         /// </summary>
-        [JsonProperty]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public int FileSize { get; set; }
 
         /// <summary>
         /// File path. Use <see cref="TelegramBotClient.GetFileAsync"/> to get the file.
         /// </summary>
-        [JsonProperty]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
         public string FilePath { get; set; }
 
         /// <summary>