+++ src/Telegram.Bot/Types/PhotoSize.cs
@@ -8,8 +8,7 @@ namespace Telegram.Bot.Types
     /// This object represents one size of a photo or a file / sticker thumbnail.
     /// </summary>
     /// <remarks>A missing thumbnail for a file (or sticker) is presented as an empty object.</remarks>
-    [JsonObject(MemberSerialization = MemberSerialization.OptIn,
-                NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
     [JsonConverter(typeof(PhotoSizeConverter))]
     public class PhotoSize : File
     {